Extract `AbstractBlackBoxContext`
